CI: reorder Rails suite library load order
move the `MyApp` Rails app out into its own `my_app.rb` file, and make
sure it gets loaded before the Rails test helpers do. those helpers will
expect an app to be in play by the time they are loaded.
